+++ src/_internals/serializer/entity-collection-serializer.ts
@@ -6,18 +6,22 @@ import type { SerializerFactory } from '.';
 const EntityCollectionSerializerFactory: SerializerFactory = () => ({
   name: 'entity-collection',
 
-  serialize: ({ value, customizerMethod, seen }): any => {
+  serialize: ({ value, customizerMethod, seen, path }): any => {
     if (value instanceof EntityCollection || (value?.__identifier__ && value.__identifier__() === 'EntityCollection')) {
-      return customizerMethod({
-        __type__: '__EntityCollection__',
-        __source__: value.source,
-        __entityName__: value.entity,
-        __context__: value.context,
-        __criteria__: value.criteria,
-        __entities__: Array.from(value),
-        __total__: value.total,
-        __aggregations__: value.aggregations,
-      }, seen);
+      return customizerMethod(
+        {
+          __type__: '__EntityCollection__',
+          __source__: value.source,
+          __entityName__: value.entity,
+          __context__: value.context,
+          __criteria__: value.criteria,
+          __entities__: Array.from(value),
+          __total__: value.total,
+          __aggregations__: value.aggregations,
+        },
+        seen,
+        path
+      );
     }
   },